refactored demo_nodes_cpp_native to use composition
