+++ demo_nodes_cpp_native/CMakeLists.txt
@@ -16,20 +16,39 @@ find_package(rclcpp REQUIRED)
 find_package(rmw REQUIRED)
 find_package(std_msgs REQUIRED)
 find_package(rmw_fastrtps_cpp QUIET)
+find_package(rclcpp_components REQUIRED)
 
-function(custom_executable target)
-  add_executable(${target} src/${target}.cpp)
-  ament_target_dependencies(${target}
-    "rclcpp"
-    "std_msgs"
-    "rmw_fastrtps_cpp")
-  install(TARGETS ${target}
-    DESTINATION lib/${PROJECT_NAME})
-endfunction()
+include_directories(include)
+
+add_library(talker_native SHARED 
+  src/talker.cpp)
+ament_target_dependencies(talker_native 
+  "rclcpp"
+  "std_msgs"
+  "rmw_fastrtps_cpp"
+  "rclcpp_components")
+rclcpp_components_register_nodes(talker_native "demo_nodes_cpp_native::Talker")
+
+add_executable(talker
+  src/talker_native_main.cpp)
+target_link_libraries(talker
+  talker_native)
+ament_target_dependencies(talker
+  "rclcpp")
+
+install(TARGETS
+  talker_native
+  ARCHIVE DESTINATION lib
+  LIBRARY DESTINATION lib
+  RUNTIME DESTINATION bin)
+
+install(TARGETS
+  talker
+  DESTINATION lib/${PROJECT_NAME})
 
 if(rmw_fastrtps_cpp_FOUND)
   find_package(rmw_fastrtps_cpp REQUIRED)
-  custom_executable(talker)
+  #custom_executable(talker)
 
   if(BUILD_TESTING)
     find_package(ament_lint_auto REQUIRED)
