+++ demo_nodes_cpp_native/include/talker.hpp
@@ -1,0 +1,41 @@
+// Copyright 2016 Open Source Robotics Foundation, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+
+#include "rclcpp/rclcpp.hpp"
+
+#include "std_msgs/msg/string.hpp"
+
+#include "rmw_fastrtps_cpp/get_participant.hpp"
+#include "rmw_fastrtps_cpp/get_publisher.hpp"
+
+
+namespace demo_nodes_cpp_native
+{
+	
+class Talker : public rclcpp::Node{
+public:
+	explicit Talker(rclcpp::NodeOptions options);
+
+	void on_timer();
+
+private: 
+	size_t count_;
+  	std::unique_ptr<std_msgs::msg::String> msg_;
+  	rclcpp::Publisher<std_msgs::msg::String>::SharedPtr pub_;
+  	rclcpp::TimerBase::SharedPtr timer_;
+
+};
+
+}
