+++ demo_nodes_cpp_native/src/talker.cpp
@@ -12,6 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+
+#include "talker.hpp"
+
 #include <iostream>
 #include <memory>
 #include <string>
@@ -21,62 +24,49 @@
 
 #include "std_msgs/msg/string.hpp"
 
-#include "rmw_fastrtps_cpp/get_participant.hpp"
-#include "rmw_fastrtps_cpp/get_publisher.hpp"
-
 using namespace std::chrono_literals;
 
-class Talker : public rclcpp::Node
+namespace demo_nodes_cpp_native
 {
-public:
-  Talker()
-  : Node("talker_native")
-  {
-    {
-      rcl_node_t * rcl_node = get_node_base_interface()->get_rcl_node_handle();
-      rmw_node_t * rmw_node = rcl_node_get_rmw_handle(rcl_node);
-      eprosima::fastrtps::Participant * p = rmw_fastrtps_cpp::get_participant(rmw_node);
-      RCLCPP_INFO(
-        this->get_logger(), "eprosima::fastrtps::Participant * %zu", reinterpret_cast<size_t>(p));
-    }
 
-    auto publish =
-      [this]() -> void
-      {
-        msg_ = std::make_unique<std_msgs::msg::String>();
-        msg_->data = "Hello World: " + std::to_string(count_++);
-        RCLCPP_INFO(this->get_logger(), "Publishing: '%s'", msg_->data.c_str());
-        pub_->publish(std::move(msg_));
-      };
-    timer_ = create_wall_timer(500ms, publish);
-    pub_ = create_publisher<std_msgs::msg::String>("chatter", 10);
-
-    {
-      rcl_publisher_t * rcl_pub = pub_->get_publisher_handle();
-      rmw_publisher_t * rmw_pub = rcl_publisher_get_rmw_handle(rcl_pub);
-      eprosima::fastrtps::Publisher * p = rmw_fastrtps_cpp::get_publisher(rmw_pub);
-      RCLCPP_INFO(
-        this->get_logger(), "eprosima::fastrtps::Publisher * %zu", reinterpret_cast<size_t>(p));
-    }
-  }
-
-private:
-  size_t count_ = 1;
-  std::unique_ptr<std_msgs::msg::String> msg_;
-  rclcpp::Publisher<std_msgs::msg::String>::SharedPtr pub_;
-  rclcpp::TimerBase::SharedPtr timer_;
-};
-
-int main(int argc, char * argv[])
+Talker::Talker(const rclcpp::NodeOptions options)
+: Node("talker", options), count_(0)
 {
-  // Force flush of the stdout buffer.
-  // This ensures a correct sync of all prints
-  // even when executed simultaneously within the launch file.
-  setvbuf(stdout, NULL, _IONBF, BUFSIZ);
 
-  rclcpp::init(argc, argv);
-  auto node = std::make_shared<Talker>();
-  rclcpp::spin(node);
-  rclcpp::shutdown();
-  return 0;
+  timer_ = create_wall_timer(500ms, std::bind(&Talker::on_timer, this));  
+
+  pub_ = create_publisher<std_msgs::msg::String>("chatter", 10);
+  
+  rcl_node_t * rcl_node = get_node_base_interface()->get_rcl_node_handle();
+  rmw_node_t * rmw_node = rcl_node_get_rmw_handle(rcl_node);
+  eprosima::fastrtps::Participant * participant = rmw_fastrtps_cpp::get_participant(rmw_node);
+  RCLCPP_INFO(
+    this->get_logger(), "eprosima::fastrtps::Participant * %zu", reinterpret_cast<size_t>(participant));
+    
+
+  rcl_publisher_t * rcl_pub = pub_->get_publisher_handle();
+  rmw_publisher_t * rmw_pub = rcl_publisher_get_rmw_handle(rcl_pub);
+  eprosima::fastrtps::Publisher * pub = rmw_fastrtps_cpp::get_publisher(rmw_pub);
+  RCLCPP_INFO(
+    this->get_logger(), "eprosima::fastrtps::Publisher * %zu", reinterpret_cast<size_t>(pub));
+
 }
+
+void Talker::on_timer()
+{
+
+  msg_ = std::make_unique<std_msgs::msg::String>();
+  msg_->data = "Hello World: " + std::to_string(count_++);
+  RCLCPP_INFO(this->get_logger(), "Publishing: '%s'", msg_->data.c_str());
+  pub_->publish(std::move(msg_));
+
+}
+
+}
+
+
+#include "rclcpp_components/register_node_macro.hpp"
+
+RCLCPP_COMPONENTS_REGISTER_NODE(demo_nodes_cpp_native::Talker)
+
+
